Thema 06 - BLAST basics/Databases 2/Deelopdracht 1/requirements.txt
astroid==2.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector==2.2.9
pylint==2.4.4
six==1.13.0
typed-ast==1.4.0
wrapt==1.11.2
